update1

+++ core_read/read.c
@@ -28,7 +28,7 @@ int do_sys_execve(struct pt_regs *ctx) {
     bpf_probe_read(&(key.pid), sizeof(key.pid), &(task->pid));
     bpf_probe_read(&(key.tgid), sizeof(key.pid), &(task->tgid));
     bpf_probe_read_str(&key.name, sizeof(key.name),
-                    (void *)(long)PT_REGS_PARM1(ctx));
+                    (void *)PT_REGS_PARM1(ctx));
     bpf_get_current_comm(&key.comm, sizeof(key.comm));
 
     bpf_perf_event_output(ctx, &counts, BPF_F_CURRENT_CPU, &key, sizeof(key));
